Porting to compile with mingw 64
Build successful with
gcc (MinGW.org GCC-6.3.0-1) 6.3.0
Copyright (C) 2016 Free Software Foundation, Inc.

+++ GUI/GUI/ScreenFader.cpp
@@ -60,5 +60,5 @@ void ScreenFader::SetColor(SDL_Color _color) {
 	SDL_DestroyTexture(fader);
 	fader = NULL;
 
-	fader = GUIHelper::SolidColourTexture(Window::Box().w, Window::Box().h, { _color.r, _color.g, _color.b, alpha });
+	fader = GUIHelper::SolidColourTexture(Window::Box().w, Window::Box().h, { _color.r, _color.g, _color.b, (Uint8)alpha });
 }
